backend/requirements.txt
PyJWT>=2.8.0
bcrypt>=4.1.0
boto3>=1.34.0
